Move PayloadUtils to request package
