+++ slack_bolt/kwargs_injection/async_utils.py
@@ -5,7 +5,7 @@ from typing import Callable, Dict, Optional, List, Any
 from slack_bolt.request.async_request import AsyncBoltRequest
 from slack_bolt.response import BoltResponse
 from .async_args import AsyncArgs
-from slack_bolt.util.payload_utils import (
+from slack_bolt.request.payload_utils import (
     to_options,
     to_shortcut,
     to_action,
