+++ slack_bolt/listener/async_listener_error_handler.py
@@ -6,7 +6,7 @@ from typing import Callable, Dict, Any, Awaitable, Optional
 from slack_bolt.request.async_request import AsyncBoltRequest
 from slack_bolt.response import BoltResponse
 
-from slack_bolt.util.payload_utils import (
+from slack_bolt.request.payload_utils import (
     to_options,
     to_shortcut,
     to_action,
